+++ HISTORY.md
@@ -5,7 +5,8 @@
 
 [Full Changelog v0.0.2-dev](https://github.com/aplazame/woocommerce/compare/v0.0.1...v0.0.2)
 
-* Fix ?
+* API exception handler
+* process_refund on gateway class
 * Full Readme.txt
 
 #### [v0.0.1](https://github.com/aplazame/woocommerce/tree/v0.0.1) (2015-09-12)